make install ok
