+++ patches/postgresql-wasm-16.4/src-port-pthread_barrier_wait.c.diff
@@ -9,14 +9,24 @@
  int
  pthread_barrier_init(pthread_barrier_t *barrier, const void *attr, int count)
  {
-@@ -33,7 +33,9 @@
- 
+@@ -75,3 +75,20 @@
+ 	pthread_mutex_destroy(&barrier->mutex);
  	return 0;
  }
--
 +#else
-+// int pthread_barrier_init(pthread_barrier_t *__restrict, const pthread_barrierattr_t *__restrict, unsigned);
++int
++pthread_barrier_init(pthread_barrier_t *__restrict barrier, const pthread_barrierattr_t *__restrict attr, unsigned count) {
++    return 0;
++}
++
++int
++pthread_barrier_wait(pthread_barrier_t *barrier) {
++    return 0;
++}
++
++int
++pthread_barrier_destroy(pthread_barrier_t *barrier) {
++    return 0;
++}
 +#endif
- int
- pthread_barrier_wait(pthread_barrier_t *barrier)
- {
++
