+++ patches/postgresql-wasm/src-interfaces-libpq-legacy-pqsignal.c.diff
@@ -4,7 +4,7 @@
   * non-ENABLE_THREAD_SAFETY builds), so the incompatibility isn't
   * troublesome for internal references.
   */
-+#if !defined(__EMSCRIPTEN__)
++#if !defined(__EMSCRIPTEN__) && !defined(__wasi__)
  pqsigfunc
  pqsignal(int signo, pqsigfunc func)
  {
@@ -12,4 +12,4 @@
  	return signal(signo, func);
  #endif
  }
-+#endif // __EMSCRIPTEN__
++#endif /* __EMSCRIPTEN__ || __wasi__ */
